% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sample.mapping.to.case.control.R
\name{convert.sample.mapping.to.case.control}
\alias{convert.sample.mapping.to.case.control}
\title{Convert sample mapping to case control from conditions}
\usage{
convert.sample.mapping.to.case.control(sample.mapping, states.case,
  states.control, out.file = NULL)
}
\arguments{
\item{sample.mapping}{Sample-condition mapping for the data set.}

\item{states.case}{Conditions to be used as case}

\item{states.control}{Conditions to be used as background}

\item{out.file}{File name to write the converted mapping}
}
\value{
a data frame containin sample mapping
}
\description{
Convert sample mapping to case control from conditions
}
\examples{
#sample.mapping = convert.sample.mapping.to.case.control(sample.mapping,
# 		    states.control = c("healthy donor"), 
#		    states.case = c("tuberculosis", "latent tuberculosis infection"))
}
