% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.z.score.R
\name{convert.z.score}
\alias{convert.z.score}
\title{Convert z score}
\usage{
convert.z.score(z, one.sided = NULL)
}
\arguments{
\item{z}{z-scores.}

\item{one.sided}{if conversion should take the sign of the z-scores (two.sided if NULL).}
}
\value{
P-values.
}
\description{
Convert z-scores to P-values.
}
\keyword{internal}
