% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.de.genes.R
\name{find.de.genes}
\alias{find.de.genes}
\title{Find differentially expressed genes}
\usage{
find.de.genes(expr, sample.mapping, states, method = "limma",
  out.file = NULL, state.background = NULL, adjust.method = "BH",
  cutoff = 0.05, functional.enrichment = NULL)
}
\arguments{
\item{expr}{expression matrix.}

\item{sample.mapping}{Sample - condition mapping.}

\item{states}{Conditions to be considered as case.}

\item{method}{Differential expression analysis method: limma (Default) | sam | welch.}

\item{out.file}{File to write output. If NULL (default) not used.}

\item{state.background}{Condition to be considered as control.}

\item{adjust.method}{Multiple hypothesis testing correction method. 
Defaults to BH.}

\item{cutoff}{Adjust p-value cutoff. Defaults to 0.05}

\item{functional.enrichment}{GO or KEGG based functional enrichment analysis}
}
\value{
data frame with results
}
\description{
Find differentially expressed genes
}
\examples{
gds.data = fetch.expression.data("GDS4966", do.log2=F, probe.conversion="Gene ID")
expr = gds.data$expr
sample.mapping = gds.data$sample.mapping
sample.mapping = convert.sample.mapping.to.case.control(sample.mapping,
 		    states.control = c("healthy donor"), 
		    states.case = c("tuberculosis", "latent tuberculosis infection"))
d = find.de.genes(expr, sample.mapping, c("case", "control"), method="limma")
}
\keyword{internal}

