% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.de.genes.sam.R
\name{find.de.genes.sam}
\alias{find.de.genes.sam}
\title{Find differentially expressed genes using SAM}
\usage{
find.de.genes.sam(expr, sample.mapping, states, out.file = NULL,
  state.background = NULL, adjust.method = "BH", cutoff = 0.05)
}
\arguments{
\item{expr}{expression matrix.}

\item{sample.mapping}{Sample - condition mapping.}

\item{states}{Conditions to be considered as case.}

\item{out.file}{File to write output. If NULL (default) not used.}

\item{state.background}{Condition to be considered as control.}

\item{adjust.method}{Multiple hypothesis testing correction method. 
Defaults to BH.}

\item{cutoff}{Adjust p-value cutoff. Defaults to 0.05}
}
\value{
Data frame with results.
}
\description{
Find differentially expressed genes using SAM
}

