% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.set.R
\name{get.data.set}
\alias{get.data.set}
\title{Get expression data set}
\usage{
get.data.set(geo.id, output.dir, is.annotation = F)
}
\arguments{
\item{geo.id}{GEO id.}

\item{output.dir}{Output directory to write / look for files.}
}
\value{
data set
}
\description{
Get expression data set
}

