% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fdr.matrix.R
\name{get.fdr.matrix}
\alias{get.fdr.matrix}
\title{Get fdr matrix}
\usage{
get.fdr.matrix(z, out.file = NULL)
}
\arguments{
\item{out.file}{Output file for writing z score matrix.}

\item{Data}{frame containing z-scores (probes vs samples).}
}
\value{
Data frame containing FDR values.
}
\description{
Calculates FDRs from z-scores for each gene in each sample.
}
\keyword{internal}

