% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.peeps.from.z.matrix.R
\name{get.peeps.from.z.matrix}
\alias{get.peeps.from.z.matrix}
\title{Get peeps from z-score matrix}
\usage{
get.peeps.from.z.matrix(z, cutoff, convert.to.pvalues)
}
\arguments{
\item{z}{Matrix containing z-scores (genes vs samples), tab separated.}

\item{cutoff}{Threshold for deciding peeps, either a z-score or adjusted p-value (if convert.to.pvalues=T).}

\item{convert.to.pvalues}{Flag to convert z-scores to p-values. 
If TRUE, the z-scores are converted to P-values which are then corrected for multiple hypothesis testing.}
}
\value{
Data frame containing sample name and geneid of genes in the peeps
}
\description{
Returns peeps for each sample in a given z-score matrix.
}
\examples{
peeps <- get.peeps.from.z.matrix(z, cutoff=0.05, convert.to.pvalues=T) 
}
