% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.platform.annotation.R
\name{get.platform.annotation}
\alias{get.platform.annotation}
\title{Get array platform annotation for probes}
\usage{
get.platform.annotation(geo.id, probe.conversion, output.dir)
}
\arguments{
\item{geo.id}{GEO id.}

\item{probe.conversion}{Id of the column used for probe gene mapping.}

\item{output.dir}{Directory to store files.}
}
\value{
Data frame containing gene mapping.
}
\description{
Get array platform annotation for probes
}

