% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.score.R
\name{get.z.score}
\alias{get.z.score}
\title{Get z score}
\usage{
get.z.score(expr, samples.background, method = "mean",
  samples.to.exclude = NULL)
}
\arguments{
\item{expr}{Expression matrix (probes vs samples).}

\item{samples.background}{Names of the background (control) samples.}

\item{method}{Method to calculate the z score, defaults to mean and sd, use median for med and mad.}

\item{samples.to.exclude}{Names of the samples to be excluded from background (for CV).}
}
\value{
Data frame containing z-scores
}
\description{
Calculates z-score for each gene in each sample
using the mean and sd over provided control samples.
}
\keyword{internal}

