% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.expression.data.R
\name{fetch.expression.data}
\alias{fetch.expression.data}
\title{Fetch expression data from GEO / or given folder}
\usage{
fetch.expression.data(geo.id, sample.mapping.column = "characteristics_ch1",
  do.log2 = NULL, probe.conversion = NULL, conversion.mapping = NULL,
  conversion.mapping.function = NULL, output.dir = paste(geo.id, "/", sep =
  ""), geo.id.sub = NULL, reprocess = NULL)
}
\arguments{
\item{geo.id}{GEO id.}

\item{do.log2}{Apply log2 transformation to the expression values. Defaults to TRUE.}

\item{probe.conversion}{Convert probe level expression to gene level using provided 
annotation label (uses platform specific annotations downloaded from GEO). 
Defaults to NULL (no conversion). In case of multiple probes, probe with absolute max
value is chosen.}

\item{conversion.mapping}{Mapping of platform specific ids to user provided ids}

\item{conversion.mapping.function}{Function to process probe name such that it matches
with the ids provided in conversion.map}

\item{output.dir}{Directory where all files will be written.}

\item{geo.id.sub}{GEO id for the sub-set (e.g., specific to the platform).}
}
\value{
A list containing 3 data frames: expression matrix, sample mapping, gene mapping.
}
\description{
Fetch expression data from GEO / or given folder
}
\examples{
gds.data = fetch.expression.data("GDS4966", do.log2=F, probe.conversion="Gene ID")
expr = gds.data$expr
sample.mapping = gds.data$sample.mapping
}

