% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fdr.matrix.R
\name{get.fdr.matrix}
\alias{get.fdr.matrix}
\title{Get fdr matrix}
\usage{
get.fdr.matrix(z, adjust.method, out.file = NULL)
}
\arguments{
\item{z}{Data frame containing z-scores (probes vs samples).}

\item{adjust.method}{P-value correction method (see p.adjust).}

\item{out.file}{Output file for writing z score matrix.}
}
\value{
Data frame containing FDR values.
}
\description{
Calculates FDRs from z-scores for each gene in each sample.
}
